testing: migrate cli unit tests to integration tests
This also refactors the CLI a bit so that it is part of the core
library, and allows greater reuse of the real CLI behavior
