+++ src/config.rs
@@ -6,7 +6,11 @@ use std::{
 use anyhow::{Context, bail};
 
 use crate::{
-    artifact::TargetTriple, channel::Channel, manifest::Manifest, toolchain::Toolchain, utils,
+    artifact::TargetTriple,
+    channel::Channel,
+    manifest::{Manifest, ManifestError},
+    toolchain::Toolchain,
+    utils,
 };
 
 /// This struct holds contextual information about the environment in which midenup/miden will
@@ -72,6 +76,21 @@ impl Config {
         };
 
         Ok(config)
+    }
+
+    /// Get the [Manifest] for locally installed toolchains
+    pub fn local_manifest(&self) -> anyhow::Result<Manifest> {
+        let local_manifest_path = self.midenup_home.join("manifest").with_extension("json");
+        let local_manifest_uri = format!(
+            "file://{}",
+            local_manifest_path.to_str().context("Couldn't convert miden directory")?,
+        );
+        match Manifest::load_from(local_manifest_uri) {
+            Ok(manifest) => Ok(manifest),
+            Err(ManifestError::Empty | ManifestError::Missing(_)) => Ok(Manifest::default()),
+            Err(err) => Err(err),
+        }
+        .context("unable to load local manifest")
     }
 
     pub fn update_opt_symlinks(
